move animations folder inside components folder
